refactor: add report target module
Introduce the report.target module.

This stage is responsible for generating the report once testing has completed.

On the guest side, it is communicated to the host that the testing has finished through tests-finished service.

On the host side, both host and guest test results are collected from the journal and parsed into the current certificate structure.

Additionally, hardcoded values have been updated to reflect the correct certificate version, as the system was previously reporting 3.0-0. The underlying logic will be revamped in a future PR; for now, this change focuses on target reorganization.

Services included in report.target:

Host:
- display-guest-logs.service
- sev-certificate-generator.service
- report-done.service

Guest:
- tests-finished.service
- report-done.service
