--- modules/logging/display-guest-logs/mkosi.extra/usr/local/lib/systemd/system-preset/10-display-guest-logs.preset
@@ -1,1 +1,0 @@
-enable display-guest-logs.service
