--- modules/logging/display-guest-logs/mkosi.extra/usr/local/lib/systemd/system/display-guest-logs.service
@@ -1,16 +1,0 @@
-[Unit]
-Description=Show the guest logs for the appropriate services in the host
-DefaultDependencies=no
-After=systemd-journal-remote.service verify-guest.service
-Wants=systemd-journal-remote.service verify-guest.service
-Requires=verify-guest.service
-
-[Service]
-Type=oneshot
-ExecStart=/usr/local/lib/scripts/display-guest-logs.sh
-StandardOutput=journal+console
-StandardError=journal+console
-LogExtraFields="SEV_VERSION=3.0-0" "SNPHOST_TEST=3.0-0"
-
-[Install]
-WantedBy=multi-user.target
