--- modules/logging/sev-certificate-generator/mkosi.extra/usr/local/lib/systemd/system-preset/10-sev-certificate-generator.preset
@@ -1,1 +1,0 @@
-enable sev-certificate-generator.service
