--- modules/logging/sev-certificate-generator/mkosi.extra/usr/local/lib/systemd/system/sev-certificate-generator.service
@@ -1,15 +1,0 @@
-[Unit]
-Description=Show the SEV certifcate
-DefaultDependencies=no
-After=network-online.service display-guest-logs.service
-Wants=network-online.service display-guest-logs.service
-
-[Service]
-Type=oneshot
-ExecStart=/usr/bin/python3 /usr/local/lib/scripts/generate_sev_certificate/sev_certificate/generate_sev_certificate.py
-StandardOutput=journal+console
-StandardError=journal+console
-LogExtraFields="SEV_REPORT=TRUE"
-
-[Install]
-WantedBy=multi-user.target
