+++ modules/report/guest/mkosi.conf
@@ -1,0 +1,3 @@
+[Include]
+Include=./tests-finished
+Include=./report-done
