+++ modules/report/host/mkosi.conf
@@ -1,0 +1,4 @@
+[Include]
+Include=./display-guest-logs
+Include=./sev-certificate-generator
+Include=./report-done
