+++ modules/report/host/sev-certificate-generator/mkosi.extra/usr/local/lib/systemd/system/sev-certificate-generator.service
@@ -1,0 +1,12 @@
+[Unit]
+Description=Generate the SEV certificate
+DefaultDependencies=no
+After=display-guest-logs.service
+Wants=display-guest-logs.service
+
+[Service]
+Type=oneshot
+ExecStart=/usr/bin/python3 /usr/local/lib/scripts/generate_sev_certificate/sev_certificate/generate_sev_certificate.py
+StandardOutput=journal+console
+StandardError=journal+console
+LogExtraFields="SEV_REPORT=TRUE"
